refactor: 정산 도메인 모델을 관심사별 패키지로 분리
- settlement, grade, promotion 모델을 역할에 맞게 재배치
- 저장소, 배치, 프레젠테이션 import 경로 정리
