--- service/settlement/src/main/java/jabaclass/settlement/domain/model/PromotionType.java
@@ -1,5 +1,0 @@
-package jabaclass.settlement.domain.model;
-
-public enum PromotionType {
-	NEW_SELLER
-}
