+++ service/settlement/src/main/java/jabaclass/settlement/domain/model/grade/SellerGrade.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.domain.model;
+package jabaclass.settlement.domain.model.grade;
 
 import java.util.UUID;
 
@@ -8,6 +8,8 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import jabaclass.settlement.domain.model.BaseEntity;
 
 @Getter
 @Entity
