+++ service/settlement/src/main/java/jabaclass/settlement/domain/model/grade/SellerGradePolicy.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.domain.model;
+package jabaclass.settlement.domain.model.grade;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -11,6 +11,8 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import jabaclass.settlement.domain.model.BaseEntity;
 
 @Getter
 @Entity
