+++ service/settlement/src/main/java/jabaclass/settlement/domain/model/grade/SellerGradeType.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.domain.model;
+package jabaclass.settlement.domain.model.grade;
 
 public enum SellerGradeType {
 	BASIC,
