+++ service/settlement/src/main/java/jabaclass/settlement/domain/model/promotion/PromotionType.java
@@ -1,0 +1,5 @@
+package jabaclass.settlement.domain.model.promotion;
+
+public enum PromotionType {
+	NEW_SELLER
+}
