+++ service/settlement/src/main/java/jabaclass/settlement/domain/model/promotion/SettlementPromotion.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.domain.model;
+package jabaclass.settlement.domain.model.promotion;
 
 import java.math.BigDecimal;
 
@@ -10,6 +10,8 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import jabaclass.settlement.domain.model.BaseEntity;
 
 @Getter
 @Entity
