+++ service/settlement/src/main/java/jabaclass/settlement/domain/model/settlement/SettlementTransferStatus.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.domain.model;
+package jabaclass.settlement.domain.model.settlement;
 
 public enum SettlementTransferStatus {
 	REQUESTED,
