+++ service/settlement/src/main/java/jabaclass/settlement/domain/repository/SellerGradePolicyRepository.java
@@ -3,7 +3,7 @@ package jabaclass.settlement.domain.repository;
 import java.math.BigDecimal;
 import java.util.Optional;
 
-import jabaclass.settlement.domain.model.SellerGradePolicy;
+import jabaclass.settlement.domain.model.grade.SellerGradePolicy;
 
 public interface SellerGradePolicyRepository {
