+++ service/settlement/src/main/java/jabaclass/settlement/domain/repository/SellerGradeRepository.java
@@ -3,7 +3,7 @@ package jabaclass.settlement.domain.repository;
 import java.util.Optional;
 import java.util.UUID;
 
-import jabaclass.settlement.domain.model.SellerGrade;
+import jabaclass.settlement.domain.model.grade.SellerGrade;
 
 public interface SellerGradeRepository {
