+++ service/settlement/src/main/java/jabaclass/settlement/domain/repository/SellerPromotionRepository.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.UUID;
 
-import jabaclass.settlement.domain.model.SellerPromotion;
+import jabaclass.settlement.domain.model.promotion.SellerPromotion;
 
 public interface SellerPromotionRepository {
