+++ service/settlement/src/main/java/jabaclass/settlement/domain/repository/SettlementHistoryRepository.java
@@ -3,7 +3,7 @@ package jabaclass.settlement.domain.repository;
 import java.util.List;
 import java.util.UUID;
 
-import jabaclass.settlement.domain.model.SettlementHistory;
+import jabaclass.settlement.domain.model.settlement.SettlementHistory;
 
 public interface SettlementHistoryRepository {
