+++ service/settlement/src/main/java/jabaclass/settlement/domain/repository/SettlementPromotionRepository.java
@@ -3,7 +3,7 @@ package jabaclass.settlement.domain.repository;
 import java.util.Optional;
 import java.util.UUID;
 
-import jabaclass.settlement.domain.model.SettlementPromotion;
+import jabaclass.settlement.domain.model.promotion.SettlementPromotion;
 
 public interface SettlementPromotionRepository {
