+++ service/settlement/src/main/java/jabaclass/settlement/domain/repository/SettlementTransferRepository.java
@@ -3,7 +3,7 @@ package jabaclass.settlement.domain.repository;
 import java.util.List;
 import java.util.UUID;
 
-import jabaclass.settlement.domain.model.SettlementTransfer;
+import jabaclass.settlement.domain.model.settlement.SettlementTransfer;
 
 public interface SettlementTransferRepository {
