+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/batch/MonthlySettlementBatchItem.java
@@ -2,8 +2,8 @@ package jabaclass.settlement.infrastructure.batch;
 
 import java.util.List;
 
-import jabaclass.settlement.domain.model.Settlement;
-import jabaclass.settlement.domain.model.SettlementTargetCalculation;
+import jabaclass.settlement.domain.model.settlement.Settlement;
+import jabaclass.settlement.domain.model.settlement.SettlementTargetCalculation;
 
 public record MonthlySettlementBatchItem(
 	Settlement settlement,
