+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/batch/SettlementTargetCalculationBatchItem.java
@@ -1,7 +1,7 @@
 package jabaclass.settlement.infrastructure.batch;
 
-import jabaclass.settlement.domain.model.SettlementTarget;
-import jabaclass.settlement.domain.model.SettlementTargetCalculation;
+import jabaclass.settlement.domain.model.settlement.SettlementTarget;
+import jabaclass.settlement.domain.model.settlement.SettlementTargetCalculation;
 
 public record SettlementTargetCalculationBatchItem(
 	SettlementTarget target,
