+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/persistence/SettlementHistoryJpaRepository.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import jabaclass.settlement.domain.model.SettlementHistory;
+import jabaclass.settlement.domain.model.settlement.SettlementHistory;
 
 public interface SettlementHistoryJpaRepository extends JpaRepository<SettlementHistory, UUID> {
 }
