+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/persistence/SettlementJpaRepository.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import jabaclass.settlement.domain.model.Settlement;
-import jabaclass.settlement.domain.model.SettlementStatus;
+import jabaclass.settlement.domain.model.settlement.Settlement;
+import jabaclass.settlement.domain.model.settlement.SettlementStatus;
 
 public interface SettlementJpaRepository extends JpaRepository<Settlement, UUID> {
