+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/persistence/SettlementTransferJpaRepository.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import jabaclass.settlement.domain.model.SettlementTransfer;
+import jabaclass.settlement.domain.model.settlement.SettlementTransfer;
 
 public interface SettlementTransferJpaRepository extends JpaRepository<SettlementTransfer, UUID> {
 }
